Ajout de getUserById afin de récupérer un utilisateur à partir de son id

+++ src/main/java/org/coiffet/tp1/Opinion/OpinionRep.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 
 public interface OpinionRep extends CrudRepository<Opinion, Long> {
 
-    Opinion getOpinionById(int id);
+    Opinion getOpinionById(long id);
 }
